requirements.txt
fastapi==0.128.3
langchain_core==1.2.9
langchain_mistralai==1.1.1
langgraph==1.0.8
psycopg==3.3.2
pydantic==2.12.5
python-dotenv==1.2.1
Requests==2.32.5
langgraph-checkpoint-postgres
uvicorn==0.34.0
